moved stores to lib
